change format print log

+++ alternate_panel.cpp
@@ -37,7 +37,7 @@ void alternate_panel::dext(Event x, double t) {
 //     NroPort 0 = controller_one
 //     NroPort 1 = generator
 //     NroPort 2 = controller two
-  int  input = *((int*)x.value);
+  double  input = *((double*)x.value);
   if (x.port == 1){
     printLog("PANEL:INPUT:GENERATOR - FLOOR %f - TIME %f \n", input, t);
   }
